Move python into a module to fix poetry install --no-root issue
