+++ bermuda_triangle/make_mines_fake.py
