+++ bermuda_triangle/templates/base.html.jinja
