+++ bermuda_triangle/templates/fake-mines.html.jinja
