+++ pyproject.toml
@@ -11,6 +11,7 @@ python = "3.11.5"
 simfile = "^2.1.1"
 gunicorn = "^21.2.0"
 flask = "^2.3.3"
+python-dotenv = "^1.0.0"
 
 
 [build-system]
